Vagrant provision, Dockerfile. Rename lastbackend exec binary name

+++ Vagrantfile
@@ -2,47 +2,72 @@
 # # vi: set ft=ruby :
 
 require 'fileutils'
-require 'open-uri'
-require 'tempfile'
-require 'yaml'
 
 Vagrant.require_version ">= 1.6.0"
 
+CLOUD_CONFIG_PATH = File.join(File.dirname(__FILE__), "/contrib/vagrant/user-data.yml")
+CONFIG = File.join(File.dirname(__FILE__), "/contrib/vagrant/config.rb")
+
+# Defaults for config options defined in CONFIG
+$num_instances = 1
+$instance_name_prefix = "core"
 $update_channel = "alpha"
+$image_version = "current"
+$enable_serial_logging = false
+$share_home = false
+$vm_gui = false
+$vm_memory = 1024
+$vm_cpus = 1
+$vb_cpuexecutioncap = 100
+$shared_folders = {}
+$forwarded_ports = {
+  2379 => 2379,
+    80 => 80
+}
+$expose_docker_tcp=2375
 
-CLUSTER_IP="10.3.0.1"
-NODE_IP = "172.17.4.99"
-NODE_MEMORY_SIZE = 1024
-USER_DATA_PATH = File.expand_path("contrib/vagrant/user-data")
-SSL_TARBALL_PATH = File.expand_path("contrib/vagrant/ssl/controller.tar")
+# Attempt to apply the deprecated environment variable NUM_INSTANCES to
+# $num_instances while allowing config.rb to override it
+if ENV["NUM_INSTANCES"].to_i > 0 && ENV["NUM_INSTANCES"]
+  $num_instances = ENV["NUM_INSTANCES"].to_i
+end
 
-system("mkdir -p ./contrib/vagrant/ssl && ./hack/ssl/init-ssl-ca ./contrib/vagrant/ssl") or abort ("failed generating SSL CA artifacts")
-system("./hack/ssl/init-ssl ./contrib/vagrant/ssl apiserver controller IP.1=#{NODE_IP},IP.2=#{CLUSTER_IP}") or abort ("failed generating SSL certificate artifacts")
-system("./hack/ssl/init-ssl ./contrib/vagrant/ssl admin kube-admin") or abort("failed generating admin SSL artifacts")
+if File.exist?(CONFIG)
+  require CONFIG
+end
+
+# Use old vb_xxx config variables when set
+def vm_gui
+  $vb_gui.nil? ? $vm_gui : $vb_gui
+end
+
+def vm_memory
+  $vb_memory.nil? ? $vm_memory : $vb_memory
+end
+
+def vm_cpus
+  $vb_cpus.nil? ? $vm_cpus : $vb_cpus
+end
 
 Vagrant.configure("2") do |config|
-  # always use Vagrant's insecure key
+  # always use Vagrants insecure key
   config.ssh.insert_key = false
+  # forward ssh agent to easily ssh into the different machines
+  config.ssh.forward_agent = true
 
   config.vm.box = "coreos-%s" % $update_channel
-  config.vm.box_version = ">= 1151.0.0"
-  config.vm.box_url = "http://%s.release.core-os.net/amd64-usr/current/coreos_production_vagrant.json" % $update_channel
+  if $image_version != "current"
+      config.vm.box_version = $image_version
+  end
+  config.vm.box_url = "https://storage.googleapis.com/%s.release.core-os.net/amd64-usr/%s/coreos_production_vagrant.json" % [$update_channel, $image_version]
 
   ["vmware_fusion", "vmware_workstation"].each do |vmware|
     config.vm.provider vmware do |v, override|
-      v.vmx['numvcpus'] = 1
-      v.vmx['memsize'] = NODE_MEMORY_SIZE
-      v.gui = false
-
-      override.vm.box_url = 'http://%s.release.core-os.net/amd64-usr/current/coreos_production_vagrant_vmware_fusion.json' % $update_channel
+      override.vm.box_url = "https://storage.googleapis.com/%s.release.core-os.net/amd64-usr/%s/coreos_production_vagrant_vmware_fusion.json" % [$update_channel, $image_version]
     end
   end
 
   config.vm.provider :virtualbox do |v|
-    v.cpus = 1
-    v.gui = false
-    v.memory = NODE_MEMORY_SIZE
-
     # On VirtualBox, we don't have guest additions or a functional vboxsf
     # in CoreOS, so tell Vagrant that so it can be smarter.
     v.check_guest_additions = false
@@ -54,22 +79,75 @@ Vagrant.configure("2") do |config|
     config.vbguest.auto_update = false
   end
 
-  config.vm.network :private_network, ip: NODE_IP
-  config.vm.network :forwarded_port, guest: 2967, host: 2967
+  (1..$num_instances).each do |i|
+    config.vm.define vm_name = "%s-%02d" % [$instance_name_prefix, i] do |config|
+      config.vm.hostname = vm_name
 
-  config.vm.synced_folder ".", "/home/core/lastbackend",
-    id: "core", :nfs => true, :mount_options => ['nolock,vers=3,udp']
+      if $enable_serial_logging
+        logdir = File.join(File.dirname(__FILE__), "log")
+        FileUtils.mkdir_p(logdir)
+
+        serialFile = File.join(logdir, "%s-serial.txt" % vm_name)
+        FileUtils.touch(serialFile)
+
+        ["vmware_fusion", "vmware_workstation"].each do |vmware|
+          config.vm.provider vmware do |v, override|
+            v.vmx["serial0.present"] = "TRUE"
+            v.vmx["serial0.fileType"] = "file"
+            v.vmx["serial0.fileName"] = serialFile
+            v.vmx["serial0.tryNoRxLoss"] = "FALSE"
+          end
+        end
+
+        config.vm.provider :virtualbox do |vb, override|
+          vb.customize ["modifyvm", :id, "--uart1", "0x3F8", "4"]
+          vb.customize ["modifyvm", :id, "--uartmode1", serialFile]
+        end
+      end
+
+      if $expose_docker_tcp
+        config.vm.network "forwarded_port", guest: 2375, host: ($expose_docker_tcp + i - 1), host_ip: "127.0.0.1", auto_correct: true
+      end
+
+      $forwarded_ports.each do |guest, host|
+        config.vm.network "forwarded_port", guest: guest, host: host, auto_correct: true
+      end
+
+      ["vmware_fusion", "vmware_workstation"].each do |vmware|
+        config.vm.provider vmware do |v|
+          v.gui = vm_gui
+          v.vmx['memsize'] = vm_memory
+          v.vmx['numvcpus'] = vm_cpus
+        end
+      end
+
+      config.vm.provider :virtualbox do |vb|
+        vb.gui = vm_gui
+        vb.memory = vm_memory
+        vb.cpus = vm_cpus
+        vb.customize ["modifyvm", :id, "--cpuexecutioncap", "#{$vb_cpuexecutioncap}"]
+      end
+
+      ip = "172.17.8.#{i+100}"
+      config.vm.network :private_network, ip: ip
+
+      # Uncomment below to enable NFS for sharing the host machine into the coreos-vagrant VM.
+      #config.vm.synced_folder ".", "/home/core/share", id: "core", :nfs => true, :mount_options => ['nolock,vers=3,udp']
+      $shared_folders.each_with_index do |(host_folder, guest_folder), index|
+        config.vm.synced_folder host_folder.to_s, guest_folder.to_s, id: "core-share%02d" % index, nfs: true, mount_options: ['nolock,vers=3,udp']
+      end
+
+      if $share_home
+        config.vm.synced_folder ENV['HOME'], ENV['HOME'], id: "home", :nfs => true, :mount_options => ['nolock,vers=3,udp']
+      end
+
+      if File.exist?(CLOUD_CONFIG_PATH)
+        config.vm.provision :file, :source => "#{CLOUD_CONFIG_PATH}", :destination => "/tmp/vagrantfile-user-data"
+        config.vm.provision :shell, :inline => "mv /tmp/vagrantfile-user-data /var/lib/coreos-vagrant/", :privileged => true
+      end
 
 
-  config.vm.provision :file, :source => SSL_TARBALL_PATH, :destination => "/tmp/ssl.tar"
-  config.vm.provision :shell, :inline => "mkdir -p /etc/kubernetes/ssl && tar -C /etc/kubernetes/ssl -xf /tmp/ssl.tar", :privileged => true
 
-  config.vm.provision :file, :source => USER_DATA_PATH, :destination => "/tmp/vagrantfile-user-data"
-  config.vm.provision :shell, :inline => "mv /tmp/vagrantfile-user-data /var/lib/coreos-vagrant/", :privileged => true
-
-  config.vm.provision "docker" do |d|
-    d.build_image "/home/core/lastbackend", args: "-t lastbackend -f /home/core/lastbackend/images/lastbackend/Dockerfile"
-    d.run "lastbackend", name: "lastbackend-daemon", args:"-v /home/core/.lastbackend/config.yml:/etc/lastbackend/config.yml -v /etc/kubernetes/ssl:/etc/lastbackend/ssl -p 2967:2967"
+    end
   end
-
 end
